internal/bench: fixed bug in benchmark filter
When the benchmark method was checked, its full name was composed
of the class name and the name without the benchmark[_] prefix,
because of this, the filter logic was broken.

For the BenchmarkSimple::benchmarkFirst method, you had to write
BenchmarkSimple::First to run just this method.

This is counterintuitive, so this behaviour has been fixed, and now
to select only this method, you need to write a logical
BenchmarkSimple::benchmarkFirst.